move model and data to main module
