+++ tsa/__init__.py
@@ -1,0 +1,1 @@
+from ._model import Model, Data
